repos/build_pipeline/requirements.txt
attrs==21.2.0; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.5.0"
aws-cdk.assets==1.127.0; python_version >= "3.6"
aws-cdk.aws-apigateway==1.127.0; python_version >= "3.6"
aws-cdk.aws-applicationautoscaling==1.127.0; python_version >= "3.6"
aws-cdk.aws-autoscaling-common==1.127.0; python_version >= "3.6"
aws-cdk.aws-autoscaling-hooktargets==1.127.0; python_version >= "3.6"
aws-cdk.aws-autoscaling==1.127.0; python_version >= "3.6"
aws-cdk.aws-certificatemanager==1.127.0; python_version >= "3.6"
aws-cdk.aws-cloudformation==1.127.0; python_version >= "3.6"
aws-cdk.aws-cloudfront==1.127.0; python_version >= "3.6"
aws-cdk.aws-cloudwatch==1.127.0; python_version >= "3.6"
aws-cdk.aws-codebuild==1.127.0; python_version >= "3.6"
aws-cdk.aws-codecommit==1.127.0; python_version >= "3.6"
aws-cdk.aws-codedeploy==1.127.0; python_version >= "3.6"
aws-cdk.aws-codeguruprofiler==1.127.0; python_version >= "3.6"
aws-cdk.aws-codepipeline-actions==1.127.0; python_version >= "3.6"
aws-cdk.aws-codepipeline==1.127.0; python_version >= "3.6"
aws-cdk.aws-codestarnotifications==1.127.0; python_version >= "3.6"
aws-cdk.aws-cognito==1.127.0; python_version >= "3.6"
aws-cdk.aws-ec2==1.127.0; python_version >= "3.6"
aws-cdk.aws-ecr-assets==1.127.0; python_version >= "3.6"
aws-cdk.aws-ecr==1.127.0; python_version >= "3.6"
aws-cdk.aws-ecs==1.127.0; python_version >= "3.6"
aws-cdk.aws-efs==1.127.0; python_version >= "3.6"
aws-cdk.aws-elasticloadbalancing==1.127.0; python_version >= "3.6"
aws-cdk.aws-elasticloadbalancingv2==1.127.0; python_version >= "3.6"
aws-cdk.aws-events-targets==1.127.0; python_version >= "3.6"
aws-cdk.aws-events==1.127.0; python_version >= "3.6"
aws-cdk.aws-globalaccelerator==1.127.0; python_version >= "3.6"
aws-cdk.aws-iam==1.127.0; python_version >= "3.6"
aws-cdk.aws-kinesis==1.127.0; python_version >= "3.6"
aws-cdk.aws-kinesisfirehose==1.127.0; python_version >= "3.6"
aws-cdk.aws-kms==1.127.0; python_version >= "3.6"
aws-cdk.aws-lambda==1.127.0; python_version >= "3.6"
aws-cdk.aws-logs==1.127.0; python_version >= "3.6"
aws-cdk.aws-route53-targets==1.127.0; python_version >= "3.6"
aws-cdk.aws-route53==1.127.0; python_version >= "3.6"
aws-cdk.aws-s3-assets==1.127.0; python_version >= "3.6"
aws-cdk.aws-s3==1.127.0; python_version >= "3.6"
aws-cdk.aws-sagemaker==1.127.0; python_version >= "3.6"
aws-cdk.aws-sam==1.127.0; python_version >= "3.6"
aws-cdk.aws-secretsmanager==1.127.0; python_version >= "3.6"
aws-cdk.aws-servicediscovery==1.127.0; python_version >= "3.6"
aws-cdk.aws-signer==1.127.0; python_version >= "3.6"
aws-cdk.aws-sns-subscriptions==1.127.0; python_version >= "3.6"
aws-cdk.aws-sns==1.127.0; python_version >= "3.6"
aws-cdk.aws-sqs==1.127.0; python_version >= "3.6"
aws-cdk.aws-ssm==1.127.0; python_version >= "3.6"
aws-cdk.aws-stepfunctions==1.127.0; python_version >= "3.6"
aws-cdk.cloud-assembly-schema==1.127.0; python_version >= "3.6"
aws-cdk.core==1.127.0; python_version >= "3.6"
aws-cdk.custom-resources==1.127.0; python_version >= "3.6"
aws-cdk.cx-api==1.127.0; python_version >= "3.6"
aws-cdk.region-info==1.127.0; python_version >= "3.6"
boto3==1.18.55; python_version >= "3.6"
botocore==1.21.55; python_version >= "3.6"
cattrs==1.8.0; python_version >= "3.7" and python_version < "4.0"
constructs==3.3.160; python_version >= "3.6"
dill==0.3.4; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.1.0"
google-pasta==0.2.0
importlib-metadata==4.8.1; python_version >= "3.6"
jmespath==0.10.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
jsii==1.38.0; python_version >= "3.6" and python_version < "4.0"
multiprocess==0.70.12.2
numpy==1.21.1; python_version >= "3.7" and python_full_version >= "3.7.1"
packaging==21.0; python_version >= "3.6"
pandas==1.3.3; python_full_version >= "3.7.1"
pathos==0.2.8
pox==0.3.0
ppft==1.6.6.4
protobuf3-to-dict==0.1.5
protobuf==3.18.1; python_version >= "3.5"
publication==0.0.3; python_version >= "3.6"
pyparsing==2.4.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
python-dateutil==2.8.2; python_version >= "3.6" and python_version < "4.0" and python_full_version >= "3.7.1" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6")
pytz==2021.3; python_full_version >= "3.7.1"
s3transfer==0.5.0; python_version >= "3.6"
sagemaker==2.59.7
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.6" and python_version < "4.0" and python_full_version >= "3.3.0"
smdebug-rulesconfig==1.0.1; python_version >= "2.7"
typing-extensions==3.10.0.2; python_version >= "3.6" and python_version < "4.0"
urllib3==1.26.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.6"
zipp==3.6.0; python_version >= "3.6"
